feat(gameloot): inject loot elements into the NFT contract
change all nft contract name.In order to achieve the effect of loot, the data structure of the
attribute has been changed. What the tokenURI method returns is no longer a complete SVG file, but
an SVG template. And the state modification of the attribute needs to be carried out in accordance
with specific rules.